requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
appdirs==1.4.4
    # via wandb
audioread==3.0.1
    # via librosa
braceexpand==0.1.7
    # via
    #   laion-clap
    #   webdataset
certifi==2023.11.17
    # via
    #   requests
    #   sentry-sdk
cffi==1.16.0
    # via soundfile
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via wandb
decorator==5.1.1
    # via librosa
docker-pycreds==0.4.0
    # via wandb
filelock==3.13.1
    # via
    #   huggingface-hub
    #   transformers
fsspec==2023.12.1
    # via huggingface-hub
ftfy==6.1.3
    # via laion-clap
gitdb==4.0.11
    # via gitpython
gitpython==3.1.40
    # via wandb
h5py==3.10.0
    # via laion-clap
huggingface-hub==0.19.4
    # via
    #   tokenizers
    #   transformers
idna==3.6
    # via requests
joblib==1.3.2
    # via
    #   librosa
    #   scikit-learn
laion-clap==1.1.2
    # via -r requirements.in
lazy-loader==0.3
    # via librosa
librosa==0.10.1
    # via
    #   laion-clap
    #   torchlibrosa
llvmlite==0.41.1
    # via
    #   laion-clap
    #   numba
msgpack==1.0.7
    # via librosa
numba==0.58.1
    # via librosa
numpy==1.26.2
    # via
    #   h5py
    #   librosa
    #   numba
    #   pandas
    #   scikit-learn
    #   scipy
    #   soxr
    #   torchlibrosa
    #   transformers
    #   webdataset
packaging==23.2
    # via
    #   huggingface-hub
    #   pooch
    #   transformers
pandas==2.1.4
    # via
    #   -r requirements.in
    #   laion-clap
platformdirs==4.1.0
    # via pooch
pooch==1.8.0
    # via librosa
progressbar==2.5
    # via laion-clap
protobuf==4.25.1
    # via wandb
psutil==5.9.6
    # via wandb
pycparser==2.21
    # via cffi
python-dateutil==2.8.2
    # via pandas
pytz==2023.3.post1
    # via pandas
pyyaml==6.0.1
    # via
    #   huggingface-hub
    #   transformers
    #   wandb
    #   webdataset
regex==2023.10.3
    # via
    #   laion-clap
    #   transformers
requests==2.31.0
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   pooch
    #   transformers
    #   wandb
safetensors==0.4.1
    # via transformers
scikit-learn==1.2.2
    # via
    #   laion-clap
    #   librosa
scipy==1.11.4
    # via
    #   laion-clap
    #   librosa
    #   scikit-learn
sentry-sdk==1.38.0
    # via wandb
setproctitle==1.3.3
    # via wandb
six==1.16.0
    # via
    #   docker-pycreds
    #   python-dateutil
smmap==5.0.1
    # via gitdb
soundfile==0.12.1
    # via
    #   laion-clap
    #   librosa
soxr==0.3.7
    # via librosa
threadpoolctl==3.2.0
    # via scikit-learn
tokenizers==0.15.0
    # via transformers
torchlibrosa==0.1.0
    # via laion-clap
tqdm==4.66.1
    # via
    #   huggingface-hub
    #   laion-clap
    #   transformers
transformers==4.35.2
    # via laion-clap
typing-extensions==4.8.0
    # via
    #   huggingface-hub
    #   librosa
tzdata==2023.3
    # via pandas
urllib3==2.1.0
    # via
    #   requests
    #   sentry-sdk
wandb==0.16.1
    # via laion-clap
wcwidth==0.2.12
    # via ftfy
webdataset==0.2.86
    # via laion-clap
wget==3.2
    # via laion-clap

# The following packages are considered to be unsafe in a requirements file:
# setuptools
